Refactor: 파일 분리 되돌리기
