+++ client/front/lib/fixerrand/fixerrandwidget/fixiscash.dart
