+++ client/front/lib/fixerrand/fixerrandwidget/fixminimap.dart
